organizadas las rutas
