+++ app.js
@@ -17,7 +17,7 @@ app.use((req, res, next) =>{
     res.status(404).render('not-found')
 })
 
-app.listen(3030, () =>{
+app.listen(3000, () =>{
     console.log(`Servidor corriendo en Puerto 3000`);
 });
