+++ controllers/mainController.js
@@ -1,10 +1,8 @@
 let express = require ('express');
-// let path = require ('path');
 const products = require ('../data/json-products.json');
 
 const mainController = {
     paginaPrincipal: (req, res) => {
-        // res.sendFile(path.resolve(__dirname, '../views/home.html'));
         res.render("home", {products : products});
     },
     search : function(req, res) {
@@ -14,8 +12,19 @@ const mainController = {
             res.render("search", {result: result});
         } catch{
             res.status(404).send('Producto no encontrado')
-        }
-        
+        }        
+    },
+
+    paginaRegister: (req, res) => {
+        res.render("register");
+    },
+
+    paginaLogin: (req, res) => {
+        res.render("login");
+    },
+
+    paginaCart: (req, res) => {
+        res.render("cart");
     }
 }
