--- routes/cart.routes.js
@@ -1,11 +1,0 @@
-let express = require ('express');
-let router = express.Router();
-let path = require ('path');
-const cartController = require('../controllers/cartController');
-
-// router para pagina de carrito
-router.get('/', cartController.paginaCart);
-
-
-
-module.exports = router;
