--- routes/register.routes.js
@@ -1,9 +1,0 @@
-let express = require ('express');
-let router = express.Router();
-let path = require ('path');
-let registerController = require ('../controllers/registerController.js');
-
-// router para pagina de registro
-router.get('/', registerController.paginaRegister);
-
-module.exports = router;
