+++ views/productDetail.ejs
@@ -14,7 +14,7 @@
                 <div class="main-img-container">
                     <h2 class="product-title"><%= product.name%></h2>
                     <%if (product.video == null) {%>
-                    <img src="/images/products/<%=product.mainImage.filename%>">
+                    <img class="mainImage" src="/images/products/<%=product.mainImage.filename%>">
                     <%}else{%>
                     <div id="player">
                         <p id="video"><%=product.video%></p>
@@ -22,9 +22,9 @@
                     <%}%>
                 </div>
                 <div class="more-img-container">
-                    <img src="/images/products/<%=product.moreImages[0].filename%>" alt="game image">
-                    <img src="/images/products/<%=product.moreImages[1].filename%>" alt="game image">
-                    <img src="/images/products/<%=product.moreImages[2].filename%>" alt="game image">
+                    <img class="clickable" data-product-index="<%= i %>" src="/images/products/<%=product.moreImages[0].filename%>" alt="game image">
+                    <img class="clickable" data-product-index="<%= i %>" src="/images/products/<%=product.moreImages[1].filename%>" alt="game image">
+                    <img class="clickable" data-product-index="<%= i %>" src="/images/products/<%=product.moreImages[2].filename%>" alt="game image">
                 </div>
             </div>
             
@@ -99,51 +99,66 @@
         </section>
     </main>
     
-        <!-- FOOTER -->
-        <%-include('partials/footer.ejs')%>
+    <!-- FOOTER -->
+    <%-include('partials/footer.ejs')%>
 
-        <script>
-           
-            let videoIdDinamico ='<%= product.video %>';
+    <script>
+        
+        let videoIdDinamico ='<%= product.video %>';
 
-            // 2. This code loads the IFrame Player API code asynchronously.
-            var tag = document.createElement('script');
-      
-            tag.src = "https://www.youtube.com/iframe_api";
-            var firstScriptTag = document.getElementsByTagName('script')[0];
-            firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
-      
-            // 3. This function creates an <iframe> (and YouTube player)
-            //    after the API code downloads.
-            var player;
-            function onYouTubeIframeAPIReady() {
-              player = new YT.Player('player', {
-                height: '360',
-                width: '640',
-                videoId: videoIdDinamico,
-                events: {
-                  'onReady': onPlayerReady,
-                  'onStateChange': onPlayerStateChange
-                }
-              });
+        // 2. This code loads the IFrame Player API code asynchronously.
+        var tag = document.createElement('script');
+    
+        tag.src = "https://www.youtube.com/iframe_api";
+        var firstScriptTag = document.getElementsByTagName('script')[0];
+        firstScriptTag.parentNode.insertBefore(tag, firstScriptTag);
+    
+        // 3. This function creates an <iframe> (and YouTube player)
+        //    after the API code downloads.
+        var player;
+        function onYouTubeIframeAPIReady() {
+            player = new YT.Player('player', {
+            height: '360',
+            width: '640',
+            videoId: videoIdDinamico,
+            events: {
+                'onReady': onPlayerReady,
+                'onStateChange': onPlayerStateChange
             }
-      
-            // 4. The API will call this function when the video player is ready.
-            function onPlayerReady(event) {
-              event.target.playVideo();
+            });
+        }
+    
+        // 4. The API will call this function when the video player is ready.
+        function onPlayerReady(event) {
+            event.target.playVideo();
+        }
+    
+        // 5. The API calls this function when the player's state changes.
+        //    The function indicates that when playing a video (state=1),
+        //    the player should play for six seconds and then stop.
+        var done = false;
+        function onPlayerStateChange(event) {
+            if (event.data == YT.PlayerState.PLAYING && !done) {
+            setTimeout(stopVideo, 6000);
+            done = true;
             }
-      
-            // 5. The API calls this function when the player's state changes.
-            //    The function indicates that when playing a video (state=1),
-            //    the player should play for six seconds and then stop.
-            var done = false;
-            function onPlayerStateChange(event) {
-              if (event.data == YT.PlayerState.PLAYING && !done) {
-                setTimeout(stopVideo, 6000);
-                done = true;
-              }
-            }
-          </script>
+        }     
+        
+    document.addEventListener('DOMContentLoaded', function () {
+        const clickableImages = document.querySelectorAll('.clickable');
+
+        clickableImages.forEach(function (clickableImage) {
+            clickableImage.addEventListener('click', function () {
+                const productIndex = this.getAttribute('data-product-index');
+                const mainImage = document.querySelector('.mainImage');
+                const tempSrc = mainImage.src;
+                mainImage.src = this.src;
+                this.src = tempSrc;
+            });
+        });
+    });
+
+</script>
 
    
 </body>
